+++ Core/Inc/Telecommunication/Packet/GPSPacket.hpp
@@ -7,7 +7,7 @@ union GPSPacket
 {
 	struct
 	{
-		uint8_t packetHeaderId;
+		uint16_t packetHeaderId;
 		uint16_t timeStamp_cs;
 		GPSValues gyroscopeValues;
 	};
